+++ scripts/generator/src/main.rs
@@ -724,6 +724,12 @@ fn get_module_bindings_config(api_version: u32) -> Vec<DirBindingsConf> {
                                  .raw_line("pub use ohos_sys_opaque_types::OH_PixelmapNative;")
 
                          },
+                         "native_node" => {
+                             builder
+                                 .blocklist_var("MAX_NODE_SCOPE_NUM")
+                                 .blocklist_var("MAX_COMPONENT_EVENT_ARG_NUM")
+                                 .raw_line("use crate::ui_input_event::ArkUI_UIInputEvent;")
+                         },
                          "native_type" => {
                              builder
                                 .raw_line("use crate::drawable_descriptor::ArkUI_DrawableDescriptor;")